pyproject.toml
[project]
name = "aim-cert-challenge"
version = "0.1.0"
description = "aim-cerification-challenge"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.14",
    "pydantic>=2.11.7",
    "python-dotenv==1.0.0",
    "python-multipart==0.0.18",
    "uvicorn>=0.34.2",
    "langgraph>=0.6.2",
    "langgraph-checkpoint>=2.1.1",
    "langgraph-prebuilt>=0.6.2",
    "langgraph-sdk>=0.2.0",
    "tavily-python>=0.7.9",
    "yfinance>=0.2.43",
    "wikipedia==1.4.0",
    "jupyter>=1.1.1",
    "langchain-experimental>=0.3.4",
    "langchain>=0.1.19",
    "langchain-cohere==0.4.4",
    "cohere>=5.12.0,<5.13.0",
    "langchain-openai>=0.3.7",
    "qdrant-client>=1.13.2",
    "rank-bm25>=0.2.2",
    "langchain-qdrant>=0.2.0",
    "ragas==0.2.10",
    "matplotlib>=3.10.3",
    "seaborn>=0.13.2",
    "rapidfuzz>=3.0.0",
]
